adds the testing for resample via linear interpolation.
LERP: Currently we fourier crop the 3d volume, then cut out a slice in 3d linear using interpolation of the FFT. This method sought to combine the steps by handling the rescaling directly during the backprojection.
The thinking being that we are actually removing a step of interpolation so this should in principle be better from an accuracy standpoint.

Importantly, it was also conceived to get around memory issues arising from padding the 3d to a power of 2 which is very problematic when > 512.

Currently: the test is skipped and the method not used in TM. The lerp has significant artifacts a high levels (> bin2) resampling and I'm not sure if its due to aliasing or something else.
The reason to keep the code is that is could be used in a fast code path when resampling at small levels.