FastFFT: part 21
Build changes:
    - GPU build flag changes to warn if double-precision is used
    - GPU build to allow NVCC try to automatically split GPU code into smaller compilation units to improve build time.
    - Force intel compiler to respect parenthesis for float math (-fprotect-parens) which is the default for clang and gnu.

Remove the options to sub-sample the TM histogram and stats arrays. This was experimental and cluttered the code, with no measured benefit.

Revert changes to StackDump::Walk b/c there are some extra warnings now in the GUI, and even though I don't think this is why, I don't want to deal with it.

Modifies the EmpiricalDistributionObject for variable precision:
    - Templated the accumulator arrays which default previously to double. Now float may be used.
        - All existing instances are explicitly set to EmpiricalDistributionObject<double> to maintain existing behavior.
    - Adds methods to calculate a running mean/variance using Welford's method (AddSampleValueForWelford)
    - Adds method to accumulate using Kahan summation to improve the accuracy, especially if floating point accumulators are used.
        - This is mutually exclusive with using Welford's method.
        - Only the AddSampleValue is different (AddSampleValueWithKahanCorrection) and the rest of the interface is the same.
    - All Getters (e.g. GetSampleSumOfSquares() are updated to consider these changes.)

Modifies TM to use parial kahan summation for the sum/sum_sqrs arrays.
    - Error correction is only tracked and applied over the n_mips stack (r/n 30 images) which provides a substantial benefit w/o i/o of tracking the errors in global memory

Modifies the TM to optionally threshold values for the stats arrays based on a running estiamte of the mean and stddev.
    - Command line flags added for dev to set both the min number of samples to include prior to thresholding and the number of standard deviations to include.
     - min-stats-counter
        - if TEST_LOCAL_NORMALIZATION is not defined or a healpix-file is not provided, this will be set to float max, which will short circuit the thresholding avoiding some overhead.
        - if TEST_LOCAL_NORMALIZATION is defined and a healpix-file is provided, this will be set to 10.0 unless --min-stats-counter is provided.
     - stats-threshold
        - defaults to 3.0f but is unused unless above conditions are met.

Adds a sum_counter array to the stats arrays in template_matching_empirical_distribution object.

Fixes a bug in TemplateMatchingDataSizer for the case where we are not resampling, but we are resizing due to FFT friendly padding.
    - Now also sets a state bool resizing_is_needed which is mutually exclusive with resampling_is_needed.

Updates the program QuickTest.cpp to allow testing GPU code.

Adds src/programs/samples/0_simple/numerical_precision.cpp to test the new numerical precision features.